requirements.txt
# Dashboard Dependencies
streamlit==1.40.0
pandas==2.2.2
plotly==5.24.1
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-ads==25.0.0
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
